+++ src/components/comics/navigation/desktop/VerticalNav.jsx
@@ -5,6 +5,7 @@ import { BOOKMARK_KEY } from "../../../..";
 import bookmarkOutline from "../../../../assets/bookies/bookies/bookie-short-line-40px.png";
 import bookmarkFilled from "../../../../assets/bookies/bookies/bookie-short-fill-40px.png";
 import { useState } from "react";
+import ComicPageAPI from "../../../api/ComicPageAPI";
 
 export default function VerticalNav(props) {
   const [bookmarkIcon, setBookmarkIcon] = useState(bookmarkOutline);
@@ -13,7 +14,9 @@ export default function VerticalNav(props) {
   );
 
   useEffect(() => {
-    localStorage.setItem(BOOKMARK_KEY, bookmarkId);
+    if (ComicPageAPI.isExistingPage(bookmarkId)) {
+      localStorage.setItem(BOOKMARK_KEY, bookmarkId);
+    }
   }, [bookmarkId]);
 
   useEffect(() => {
